requirements.txt
# requirements.txt
oss2==2.19.1
python-dotenv==1.1.0